chore: clean up mess
