+++ zparse/src/test_utils.rs
@@ -1,18 +1,15 @@
 mod comparison;
 mod data;
-mod fixtures;
 mod helpers;
 
-pub use comparison::*;
-pub use data::*;
-pub use fixtures::*;
-pub use helpers::*;
+pub use comparison::{assert_values_equal, compare_values};
+pub use data::TestData;
+pub use helpers::tmp_file_path;
 
 // Re-export common test types/traits
 pub use crate::{
-    common::value_compare::values_equal,
     converter::Converter,
-    error::{
+    error::{ConversionError, 
         FormatError, LexicalError, ParseError, ParseErrorKind, Result, SecurityError,
         SemanticError, SyntaxError,
     },
@@ -20,7 +17,9 @@ pub use crate::{
     parse_file,
     parser::{
         config::{ParserConfig, DEFAULT_MAX_DEPTH, DEFAULT_MAX_OBJECT_ENTRIES, DEFAULT_MAX_SIZE},
-        JsonParser, TomlParser, Value,
+        json::JsonParser,
+        toml::TomlParser,
+        value::{values_equal, Value},
     },
     utils::{format_json, format_toml, parse_json, parse_toml, read_file, write_file},
 };
